refactor: move dp code to own module
